refactor: 컴포넌트 모듈화 구조 개선
